Refine view, add all icons
