+++ src/components/Asset/assets.scss
@@ -1,0 +1,173 @@
+
+.asset-container{
+  background-color: $asset-target-color;
+  width: 50%;
+  padding: 10px 40px;
+  border-top-right-radius: $main-window-border-radius;
+
+  hr {
+    width: 90%;
+    background-color: rgba(#919191, 0.2);
+    height: 1px;
+    margin: -2px auto 0;
+  }
+
+  .input-section{
+    display: flex;
+    width: 90%;
+    margin: 0 auto;
+    padding-top: 15px;
+
+    .dropdown{
+      position: relative;
+      margin-top: 0.25em;
+
+      width: 40%;
+      line-height: 2em;
+      font-size: 1em;
+
+      background-color: transparent;
+      color: white;
+      outline: none;
+
+      hr {
+        width: 100%;
+      }
+
+      .selected {
+        display: flex;
+        cursor: pointer;
+        text-align: left;
+
+        .symbol-dropdown {
+          color: $secondary-color;
+          margin-top: 10px;
+        }
+
+        .image-currency {
+          width: 30px;
+          height: 30px;
+          margin: 5px 8px 0;
+        }
+
+        p {
+          width: fit-content;
+          margin-bottom: 0;
+        }
+      }
+
+      .dropdown-window {
+        position: absolute;
+        width: 250%;
+
+        border-bottom-left-radius: 20px;
+        border-bottom-right-radius: 20px;
+
+        z-index: 1;
+        cursor: pointer;
+        backdrop-filter: blur(30px);
+        background-color: $dropdown-menu-color-transparent;
+        overflow: hidden;
+
+        .items-search-box {
+          position: relative;
+          justify-content: space-around;
+          padding: 10px 5px 10px;
+
+          .input-search {
+            width: 90%;
+            background-color: rgba(#e3e3e3, 0.3);
+            color: white;
+            border: white solid 1px;
+            padding: 0 20px;
+            border-radius: 3px;
+            outline: none;
+          }
+
+          .symbol-search {
+            position: absolute;
+            top: 1em;
+            right: 3em;
+            font-size: 1em;
+          }
+        }
+
+        .items-wrapper {
+          max-height: 15em;
+          overflow-y: auto;
+          scrollbar-width: none;
+
+          .items{
+            text-align: left;
+            .image-currency {
+              width: 20px;
+              height: auto;
+              margin: 0 20px 0;
+            }
+          }
+
+          .items:hover {
+            background-color: $dropdown-menu-color-transparent-highlight;
+          }
+        }
+
+        .items-wrapper::-webkit-scrollbar {
+          display: none;
+        }
+      }
+
+      .hideDropDown {
+        display: none;
+      }
+    }
+
+    .asset-amount {
+      width: 60%;
+      background-color: transparent;
+      border: none;
+      font-size: 1.5em;
+      text-align: center;
+      color: white;
+      outline: none;
+    }
+  }
+
+  .overview {
+    height: 8em;
+    padding-top: 2.5em;
+
+    .text-fiat-value {
+      color: #919191;
+      width: 100%;
+
+      p {
+        font-size: 3em;
+        line-height: 1em;
+      }
+    }
+
+    .text-equivalence{
+      display: block;
+
+      color: rgba($secondary-color, 0.5);
+      font-size: 0.7em;
+      width: 100%;
+    }
+  }
+}
+
+.originAsset {
+  background-color: $asset-origin-color;
+  border-top-left-radius: $main-window-border-radius;
+  border-top-right-radius: 0;
+}
+
+@supports (-moz-appearance:none) {
+  .asset-container .input-section .dropdown .dropdown-window  {
+    background-color: $dropdown-menu-color-opaque;
+
+    .items-wrapper .items:hover {
+      background-color: $dropdown-menu-color-opaque-highlight;
+    }
+  }
+}
